Turns out there was more to solve to get it compiling

+++ aiemuc/system.c
@@ -1,7 +1,7 @@
 #include "system.h"
 //#include "device.h"
 
-uint8_t system_device_register(system_8bit_t* system, void* device, device_descriptor_t* desc, uint16_t base, uint16_t size) {
+uint8_t system_device_register(system_8bit_t* system, void* device, device_descriptor_t* desc, uint16_t base, unsigned int size) {
     if (system->device_count >= 16) return 0xFF;
     uint8_t id = system->device_count++;
     device_entry_t* entry = &system->devices[id];
